+++ src/rfu_union_tool.c
@@ -487,7 +487,7 @@ static u32 RfuUnionGroupMemberIsInvisible(u32 group, u32 member)
     return RfuUnionObjectIsInvisible(5 * group + member - 0x38);
 }
 
-static void SpawnGroupMember(u32 groupNo, u32 memberNo, u8 direction, struct GFtgtGname * gname)
+static void SpawnGroupMember(u32 groupNo, u32 memberNo, u16 direction, struct GFtgtGname * gname)
 {
     s32 x, y;
     s32 objId = 5 * groupNo + memberNo;
